migrations and localization fixes

+++ database/migrations/2021_04_16_110002_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->id();
             $table->bigInteger('user_id')->unsigned();
             $table->string('transaction_id');
-            $table->integer('total')->default(0);
+            $table->float('total')->default(0);
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users');
         });
